Update repo package publishing
The DEB and RPM repo packages must be excluded when synchronizing the repositories
to avoid deleting them. The repo packages are handled in the separate workflow.

The PR also updates the uploading of the repo packages to use "aws s3 cp" instead
of "aws s3 sync", and to allow deploying new versions without updating the package
name.